src/requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.12.1
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
datasets==4.6.0
debugpy==1.8.20
decorator==5.2.1
dill==0.4.0
exceptiongroup==1.3.1
executing==2.2.1
filelock==3.24.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.4.1
idna==3.11
ipykernel==7.2.0
ipython==8.38.0
ipywidgets==8.1.8
jedi==0.19.2
jinja2==3.1.6
jupyter-client==8.8.0
jupyter-core==5.9.1
jupyterlab-widgets==3.0.16
kiwisolver==1.4.9
markdown==3.10.2
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
packaging==26.0
pandas==2.3.3
parso==0.8.6
pillow==12.0.0
platformdirs==4.9.2
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
pure-eval==0.2.3
pyarrow==23.0.1
pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
requests==2.32.5
rich==14.3.3
scipy==1.15.3
seaborn==0.13.2
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
torch==2.10.0+cu128
torchvision==0.25.0+cu128
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typer-slim==0.24.0
typing-extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
werkzeug==3.1.7
widgetsnbextension==4.0.15
xxhash==3.6.0
yarl==1.22.0
sagemaker